Expose clipboard shortcut in command settings
