+++ docs/features/file-search.md
@@ -134,9 +134,9 @@ guards entry into `.fileSearch`, so neither a stale selected command nor the glo
 the screen after the feature is disabled. Disabling cancels the session and returns an open File Search
 screen to the launcher without changing palette visibility.
 
-Search Files is the only built-in command with a `HotKeyAction` of its own, so `AppEntry.hotKeyAction`
-returns non-nil for it alone. Two consequences follow: its launcher row prints the chord as a keycap,
-and the Search Files row in Settings ▸ Commands carries the same recorder as the one in this pane —
+Search Files and Clipboard History are the built-in commands with a `HotKeyAction` of their own, so
+`AppEntry.hotKeyAction` answers for both. Two consequences follow: their launcher rows print a bound
+chord as a keycap, and each row in Settings ▸ Commands carries the same recorder as its feature pane —
 one binding reachable from two places, not two settings.
 
 Launcher visibility is `VisibilityStore`'s, keyed on the entry's `preferenceKey`, which is why the
